[Support] Removed unnecessary recipients adapter and refactored names
