+++ src/Sylius/Bundle/SupportBundle/Resources/config/recipients.yml
@@ -1,3 +1,3 @@
 parameters:
-    sylius.support.notification.recipients:
+    sylius.support.recipients:
         - 'example@example.com'
