+++ src/Sylius/Bundle/SupportBundle/Resources/config/services.xml
@@ -17,17 +17,12 @@
                                http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <parameters>
-        <parameter key="sylius.adapter.support_notification_recipient.class">Sylius\Component\Support\Adapter\ConfigNotificationRecipientAdapter</parameter>
-        <parameter key="sylius.provider.support_notification_recipient.class">Sylius\Component\Support\Provider\NotificationRecipientProvider</parameter>
+        <parameter key="sylius.provider.support.array_recipients.class">Sylius\Component\Support\Provider\ArrayRecipientsProvider</parameter>
     </parameters>
 
     <services>
-        <service id="sylius.adapter.support_notification_recipient" class="%sylius.adapter.support_notification_recipient.class%">
-            <argument key="emails">%sylius.support.notification.recipients%</argument>
-        </service>
-
-        <service id="sylius.provider.support_notification_recipient" class="%sylius.provider.support_notification_recipient.class%">
-            <argument type="service" id="sylius.adapter.support_notification_recipient" />
+        <service id="sylius.provider.support.array_recipients" class="%sylius.provider.support.array_recipients.class%">
+            <argument key="emails">%sylius.support.recipients%</argument>
         </service>
     </services>
