--- src/Sylius/Component/Support/Provider/NotificationRecipientProviderInterface.php
@@ -1,23 +1,0 @@
-<?php
-
-/*
- * This file is part of the Lakion package.
- *
- * (c) Lakion
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
- 
-namespace Sylius\Component\Support\Provider;
-
-/**
- * @author Grzegorz Sadowski <grzegorz.sadowski@lakion.com>
- */
-interface NotificationRecipientProviderInterface
-{
-    /**
-     * @return array
-     */
-    public function getEmails();
-}
